requirements.txt
fastai>=2.0.0
torch-summary
matplotlib
